Implement a more general "by rec" allowing to use
an arbitrary term as the recursion argument, but restricting
what definitions can be made when parameters are involved (examples
in the test-suite). Now lexicographic ordering should be readily
available. When doing recursion on inductive families, one has to
explicitely use packing when calling the homogeneous relation on the
packed type.