Fix a bug in specialize_eqns not doing enough unification
We now allow convertible constructor parameters in the equalities we simplify.
This limitation was resulting in simplify_IH_hyps failing to simplify when it
could. Also reworked the way specialize_eqns performs unifications to
avoid (apparently) exponential behavior of unification in presence of
let-bound evars when we know the solutions to the evars cannot come from
those lets (substitute instead of keeping them in the context).

Fix examples accordingly, and cleanup poly_bocquet to use more of funelim.