RoseTree example: Use a more general map function
The `map_In` that I use here has (I believe) the most general type that
still allows the reasoning required to make the nested examples go
through. Note that it does not depend on the actual size function, and
would work equally well in a proof with a well-founded relation instead
of a measure.