Workaround to a completely crazy handling of universe contexts.
Since the Program (and thus Equations) API is relying on a bunch of hacks
to handle universe constraints, this results in Equation haphazardly
performing evarmap surgery to make things go through. Unfortunately
with the arrival of sort qualities, many hacks are not working anymore
as the plugin sends unbound qualities to the kernel, which is now an
anomaly.

Instead of trying to fix the utterly nonsensical evarmap threading, which
would be a daunting task of its own, we hack around the Program code to
ensure we never define terms with unresolved qualities. Basically,
we intercept terms sent to the kernel and fix their relevance by hand.
This is not very satisfying but at least it is a quick hack to make progress
on the SProp front of Coq.