Fix test-suite and examples, now requiring the stdlib files.
Issues with import order appeared. A new problem is that due to the use of a custom clos_trans definition, previous scripts which were automatically solving obligations with the "program_simpl" tactic now require to use `equations_simpl` instead. We put explicit "Local Obligation Tactic := CoreTactics.equations_simpl` directives to fix this (which is generally a good idea to document the automation being used).